fix: resolve dtype mismatch in LAION and missing method in MPS
- Cast LAION image features to dynamically match the linear head's dtype to prevent Float/Half collision
- Implement _process_image helper in MPS evaluator to handle PIL to FP16 tensor conversion
- Ensure robust pipeline execution for all 8 evaluators during A/B testing

+++ src/evaluators/aesthetic_eval.py
@@ -85,7 +85,7 @@ class AestheticEvaluator(BaseEvaluator):
         self.mlp = MLP(768)  # ViT-L/14 output dimension is 768
         state_dict = torch.load(self.mlp_path, map_location=self.device)
         self.mlp.load_state_dict(state_dict)
-        self.mlp.half().to(self.device)
+        self.mlp.to(self.device)
         self.mlp.eval()
         
         logger.info(f"{self.evaluator_name} loaded successfully.")
